Feat: add elem to /game page
