+++ public/game.css
@@ -5,6 +5,14 @@
     height: 800px;
 }
 
+#opponentUserInfo {
+    flex-grow: 1;
+}
+
+#myUserInfo {
+    flex-grow: 1;
+}
+
 .userInfoArea {
     display: flex;
     flex-direction: column;
@@ -31,10 +39,28 @@
     flex-grow: 1;
 }
 
+.chatArea {
+    display: flex;
+    flex-direction: column;
+    flex-grow: 5;
+}
+
+#gameInfo {
+    flex-grow: 1;
+}
+
+#chatLog {
+    flex-grow: 8;
+}
+
+#chatInput {
+    flex-grow: 1;
+}
+
 .elem {
     border: solid 1px black;
-    flex-grow: 1;
     border-radius: 10px;
     padding: 5px;
     margin: 5px
 }
+
